refactor encrypt method and remove sub package
